Rework cooldowns, add gambling multiplier
Cooldown reducers will now only be applied upon the cooldown being set,
not when being read. Gambling multiplier can be acquired by getting
every gambling achievement.

+++ Extensions/DateTimeOffsetExt.cs
@@ -2,7 +2,7 @@
 namespace RRBot.Extensions;
 public static class DateTimeOffsetExt
 {
-    public static long ToUnixTimeSeconds(this DateTimeOffset ts, double addSecs)
+    public static long ToUnixTimeSeconds(this DateTimeOffset ts, long addSecs)
     {
         TimeSpan epoch = DateTime.UtcNow - new DateTime(1970, 1, 1, 0, 0, 0);
         return (long)(epoch.TotalSeconds + addSecs);
